+++ test/js/testParse.js
@@ -137,6 +137,40 @@ describe('parse', () => {
 
       assert.deepStrictEqual(actual, expected);
     });
+
+    it('accepts non-standard font weights', () => {
+      const font = {
+        variants: ['heavy'],
+      };
+      const data = {
+        formats: 'woff',
+        heavy: {
+          path: 'font/path',
+          svgid: 'svgid',
+        },
+      };
+      const actual = parse.font(font, data);
+      const expected = [
+        {
+          variant: 'heavy',
+          isLocal: true,
+          hasEmbedded: false,
+          family: undefined,
+          formats: {
+            woff: {
+              src: path.normalize('font/path.woff'),
+              dest: 'assets/fonts/font/path.woff',
+            },
+          },
+          style: 'normal',
+          svgid: 'svgid',
+          weight: 900,
+          local: undefined,
+        },
+      ];
+
+      assert.deepStrictEqual(actual, expected);
+    });
   });
 
   describe('customCSS', () => {